Updated FR to use the new Reports module, proflie directive
The frontend code has been updated to use the new permissions as
outlined in the reports module. This mainly affects the admin /
moderation section. To help move this along, the profile view has been
turned into a directive so that it can be set programatically through
the controller rather than through ui-router.