No longer link to 'deleted' on all pages
When a post or user was deleted and they were the last person to post in
a board, 'deleted' would show up as a link. This has been changed so
that deleted is still there but is no longer clickable.

Pages affected:
- Boards (homepage)
- threads
- recent threads
- posted in threads
- watchlist
- board children in threads

The recent threads section was also revamped to work more similarly to
how threads currently work. Also all reply/views columns have a default
value of 0 if they don't have a value.