Bug Fix: Editor hides last post
Fixed a bug where the editor hides the last post on the page. To fix
this, the resizeable directive keeps track of how much the editor grows
in the y direction. It then adds that height to a spacer on the posts
page. This effectively pushes the content above the top of the editor.