Support for board/category postable_by, writeAccess
Adds support for setting the postable_by flag on boards and categories
in the admin/management/boards view. This allows an admin to set whether
a user can post to any board or category. As with the viewable_by flag,
any parent with this value set, the children of those parents will also
abide by the postable_by of the parent.

The threads and posts page were both updated to make sure of the
writeAccess property that is now passed to the frontend. This informs
the user of whether they are allowed to post/create a thread in the view
they are currently in.