Bug Fix: target="_blank" as an attack vector
This commit address the target="_blank" being a possible phishing attack
vector. This attack is being mitigated by running a jquery script that
adds rel="noopener noreferrer" to all links with a target="_blank"
attribute. This script was added to the post-processing directive so
anywhere posts are being processed, the script should be run as well.

For more details: https://mathiasbynens.github.io/rel-noopener/