Patroller and newbie support / Patroller page
The admin roles managements editor was updated to support the new
priority condition for all module types.

The permission logic on the posts page was also updated to handle the
new priority condition. Some of the permission logic was also updated to
not short circuit so early and to be more readable. The posts page was
also updated to display the role_name instead of the default 'HERO'
text.

The post-processing directive was refactored to remove all unused vars
and to remove a errant console.log.

A new patrol page was also added that lists all the most recent posts by
newbies and with the controls to delete/report/update the newbie's post.
This makes use of the new patroller plugin created in epochtalk.