Post Locking
The frontend was updated to support post locking on the post view. This
required adding a new permission method to check whether a user is
allowed to lock a post but also included an update to the delete modal
so that an admin/moderator type role can quick lock a post while it's
being deleted. The delete resource was also updated to handle the new
"locked" query param for this quick lock.

The SCSS was also updated to highlight the lock icon when a post is
locked.