Bug Fix: back button on admin pages
There was an issue with the back button on the leading admin pages,
(settings/moderation/management) where if the url was something like
/admin/settings it would redirect to the proper page. This redirect
would be push to the history and the user would get 'locked' into a page
they can't back out of.

This has been fixed by not pushing any of the redirect state on to
history. A side effect of this is that going to any of the redirect
routes (ie. /admin/settings) no longer go to a login page if the user is
not logged in. Those routes will now redirect to the home page.