Bug Fix: back button on profile post to admin users
Fixed an issue with using the back button to go from the profile posts
page back to the admin users page. The issue lies in the pullPage method
in the profile posts controller that would inject an extra
$state.go('.').

This caused ui-router to inject an extra page change on each page
change.