Bug Fix: Update to using new parser pattern
The messaging view was still using the old way of parsing by referencing
the epochtalk-bbcode-parser rather than using the list of parsers
available from window.parsers. This has been updated and the message
body is now bound to the partial output in the message list using
ng-bind-html rather than handlebars.