Message user from their profile page
This commit adds the ability to start a conversation with a user from
their profile page. It basically takes the same model and controller
code from the messaging page and tweaks it a little. It binds the
receiver to the id of the profile page and also redirects to the
messaging view after the conversation is started.