Bug fix: Paging profile posts
Fixed a bug where the posts on a user's profile page could no longer by
paged. This error started when the profile page was turned into a
directive. A regression occured where the child state profile.posts was
no longer being recreated on location change. This commit addresses the
issue by changing the way the profile page is being called in ui-sref.
Instead of going to the profile state directly, the ui-sref now
references the child state, profile.posts which will instantiate the
parent state automatically.

Many files were updated due to the change in API. This commit captures
all those changes in the core modules but the extra modules will have to
be updated separately.