Overhauled Notifications and Websocket services
A Notification service has been created to handle the management of
notifications for the user. It was originally held in the header
controller but has been moved out into it's own file. The same can be
said for Websockets, although websockets did originally have a service
file. The new Websockets implementation is a more comprehensive approach
with all the websocket code residing in the service file and only
exposing an API to the socket connection rather than exposing the socket
connection itself.

The patrol, posts, profile and messages views were all overhauled to use
the new Websocket implementation and all the views except for messages,
now display whether a user is online or not.