Safari Bug Fix: saving profile without a dob
If a user tries to update their profile and doesn't include a dob, the
update fails and the loading spinner continues forever. The reason is
because of a JS failure trying to create a new Date from undefined. This
commit fixes it by only creating a new date if a dob value is found.