Redirect to front page if logging out from private board
Added an extra error handler that checks if the user has reached a 404
error on the 'threads.data' or board view, and redirects to the front
page. The only reason a user would get a 404 on a board view is if the
board doesn't exist or if it is private. In either scenario, they are
redirected to the front page.