Legal Page and static footer
A new admin view to edit the contents of the legal page was added to the
settings section of the admin view. All the plumbing for this page was
done so that this page is only visible to users with the
adminAccess.settings.legal permission. The Settings header was also
updated to include a new tab for legal. The correct resources were also
created for the three routes and now live in AdminLegal.

The roles view was also updated to account for the three new permissions
that guard the viewing, updating, and resetting of the legal text. The
permission to view the legal page was also added to the roles view.

The footer was removed in favor of a static footer that has already been
added in the epochtalk repo.