Updated the profile posts controller to match pageByUser
The pageByUser route was updated to no longer take in a field and to
default the sorting direction to true (descending). The frontend was
updated to match this decision by mainly taking away the extra sortable
icon in thread and to update the pullPage and locationChangeSuccess
methods to match the changes in the route.