Updates to ImageUploader: refactoring
The image uploader directive was refactored because there was similar
code between two functions which has now been pulled into one function.
Also, a if/then check was also added to ensure that images are being
uploaded and if no images are found, then a warning is issued to the
user. This fixes an issue where if a file other than a image was chosen,
it may be uploaded to s3/local and also fixes the issue where if the
file was rejected, the loading bar still appeared.