Support avatar url substitution
The frontend was updated to support avatar url substitution. This
manifests as an update to the profile avatar and header avatar section
by adding the image-loader directives to those sections. The
image-loader directive was also updated to run again if the element's
src is ever changed.

CSS changes were also made to ensure that the loader gif is shown when
the image is been waited on.