Bugfix: slide toggle clips content on resize
The slide toggle directive that controls showing and hiding content has
an issue where if the screen is resized, content that is now longer than
it was previously, could be clipped.

The fix to this was to add an event listener to the page resize event
and recalculate the true height of each element and setting the
appropriate max-height css value after the resize.

The resize event is also debounced so that the event should only fire
once per window resize.