Always show compose tab in editor on init
Fixed an issue where if the user was on the preview tab and left the
editor for any reason, the next time the editor was opened, it would
still be on the preview tab. The quick fix for this was to add a line to
the init function to force the editor to open with the compose tab
instead.

This was also added to the image insert function so that when the user
chooses to insert an image, the compose tab was come back into focus. A
few other UI hints were added to bolster this affect. Most notable is
the image header flashes the primary color and the text 'Add to Editor'
is changed to 'Added'.