Portal System frontend implementation
Updated the '/' route to be able to fork between the 'portal' state or
the 'boards' state based on a variable passed to the public/index.html
page.

Added a 'portal' state that queries the /api/portal routes to display a
feed of the most recent threads and their last post on the portal page,
along with a list of the most recent threads globally.

Support for enabling and setting the board id for the portal in the
admin settings section.

Added a tab on the roles modal that can set whether a role can view the
portal or not.

Quick fix to the humanDate filter. at -> 'at'

Updated all the ui-sref linking to the boards state or point to home
instead and setting a 'home' state that forks to either 'portal' or
'boards' based on the public/index.js vars.