1.0.4: Invitations
Frontend has been updated to support invitations. The user drop down has
sports a link that opens the invite model where an admin can enter the
email of the person they want to invite. The admin / management page now
has a sub section for invitations where admins can view, resend, and
remove invitations as well as another area to send invitations from.

This involved the creation of invitation state and an invitation modal
on the angular side.

A new /join state was added to handle registering users from an
invitation email.