1.0.6: Portal Bug Fix
Fixed a bug where if the portal is being used and users are viewing the
portal, but then the portal is turned off, the portal page returns an
error. This error is now elegantly handled and the users are instead
redirected to the boards view.

The url scheme for boards and portal has been combined such that '/'
just goes to the right resource.