1.0.7: Saas mode in admin/settings/general
Rearranged the sections in the admin/settings/general page such that if
the images and emailer section were to be removed, the flow of the page
would still look aright.

The images and emailer sections were also given an ng-if that checked to
see if saasMode was enabled or not. If enabled, those sections are not
shown.