1.0.11: Set Moderator Directive
Created a directive where given a board, the user with the correct
permissions can add and/or remove moderators to that board. This
includes typing in a username and getting a list of matching usernames
to be added to the list of moderators to this board. It also includes
the running list of current moderators and a control to remove them as
needed.

The code for the directive was taken from the modal that does the same
action in the admin/management/boards view. It was re-written to be
smaller and more efficient. This directive now replaces that modal in
that view as well as being added to the main boards view so that this
action can be done by an admin from any board.