1.0.13: User Preferences
The frontend was update with the ability to save the number of posts
shown on a page, the number of threads shown on a page, and the
categories to hide on refresh. These values are stored on both the
browser (using localStorage) and on the server in a table.

A preference service, PreferenceSvc, was created to handle the state of
the preference data. A new 'preferences' route was also added to the
User resource to pull this data down from the server. Any update to this
data is done through the usual User.update method.

This service in also tied into the Auth service. Upon login or
authenticate, the PreferencesSvc.pullPreferences method is called that
will pull the latest preference data from the server and storage in
locally. This data is also open to the rest of the system via the
PreferencesSvc.preferences property.

The Profile View was also updated with an extra modal that now allows
the user to set the number of posts and threads per page.

The posts/index.js and controller were also updated to use any limit
found in the url first, then default to the PreferenceSvc, and finally
the default of 25 if nothing else is found.

The same was done for threads/index.js and its controller but the
posts_per_page is also used to calculate the number of pages per
threads.

Finally the boards view was overhauled to pull in the preference data on
bootstrap and map through the categories to see which one should be
shown. The ng-click on the chervon was also updated to call a method
that not only toggle the visibility of the category but also updated the
local preferences and sends a request to update the server's copy as
well.